WIP #2 - Remove support for Castle.Core 3.2.0
This is because in all 3.x versions of Castle.Core, the IProxyGenerator
interface is missing.  The proxy generator class does not implement any
interfaces.  This would require us to depend on the concretion (bad).

In theory it could be possible to create a separate build of AutoLazy
to support CC 3.x, using a compiler constant to switch the interface
for the concrete class in that build config, but there's no call for that
right now, just to support an obsolete version.